+++ demod/mod/win-script/rs41mod_live.bat
@@ -1,1 +1,1 @@
-sox.exe -q -t waveaudio "CABLE Output" -t wav - 2>nul | rs41mod.exe --ecc4 -r | tee -a rs41_raw_%date:~-4,4%%date:~-7,2%%date:~-10,2%-%time:~-11,2%%time:~-8,2%%time:~-5,2%.txt | rs41mod.exe --ecc4 --ptu2 --dewp -vx --rawhex | tee -a rs41_data_%date:~-4,4%%date:~-7,2%%date:~-10,2%-%time:~-11,2%%time:~-8,2%%time:~-5,2%.txt
+sox.exe -q -t waveaudio "CABLE Output" -t wav - 2>nul | rs41mod.exe --ecc4 -r | tee -a "rs41_raw_%date:~-4,4%%date:~-7,2%%date:~-10,2%-%time:~-11,2%%time:~-8,2%%time:~-5,2%.txt" | rs41mod.exe --ecc4 --ptu2 --dewp -vx --aux --rawhex | tee -a "rs41_data_%date:~-4,4%%date:~-7,2%%date:~-10,2%-%time:~-11,2%%time:~-8,2%%time:~-5,2%.txt"
